python-enterprise-platform/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
flask==3.0.0
flask-restful==0.3.10

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2
tortoise-orm==0.20.0
asyncpg==0.29.0

# 缓存
redis==5.0.1
hiredis==2.3.2
python-memcached==1.59

# 消息队列
celery==5.3.6
kombu==5.3.5
kafka-python==2.0.2
pika==1.3.2

# 认证授权
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# 配置管理
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
dynaconf==3.2.4

# 日志
structlog==24.1.0
python-json-logger==2.0.7
loguru==0.7.2

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# 数据验证
marshmallow==3.20.1

# 任务调度
apscheduler==3.10.4

# 监控
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# 代码质量
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# 工具库
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# 序列化
orjson==3.9.10
msgpack==1.0.7

# 限流
slowapi==0.1.9

# 依赖注入
dependency-injector==4.41.0

# Elasticsearch
elasticsearch==8.11.1

# 文档
mkdocs==1.5.3
mkdocs-material==9.5.3
